+++ app/Livewire/Gedung/GedungIndex.php
@@ -11,6 +11,14 @@ class GedungIndex extends Component
     #[Layout('layouts.app')]
 
     public function delete($id)  {
+        $gedung = gedung::findOrFail($id);
+
+        // Periksa apakah gedung memiliki relasi di tabel Penempatan
+        if ($gedung->ruangans()->exists()) {
+            // Jika memiliki relasi, tampilkan pesan gagal dihapus
+            return redirect()->back()->with('gagal', 'gedung tidak dapat dihapus karena sudah memiliki ruangan, silakan cek di data ruangan');
+        }
+
         gedung::find($id)->delete();
         session()->flash('sukses', 'Data Deleted');
     }
